+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### New Features
 
+* Added support for the `KHR_node_visibility` extension.
+
 * Added new `NODE_SKINNED_MESH_PARENT_TRANSFORMS` validation warning.
 
 ### Changes